fix compiler waring (regarding size_t and 64bit architecture)
git-svn-id: svn://svn.minisip.org/minisip/trunk@2195 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libminisip/source/sdp/SdpHeaderA.cxx
@@ -62,7 +62,7 @@ string SdpHeaderA::getString(){
 }
 
 string SdpHeaderA::getAttributeType(){
-	uint32_t pos;
+	size_t pos;
 	
 	if(( pos = attributes.find( ":" )) == string::npos )
 		return "property";
